Enhance Active Directory Automation
Added the functionality to be able to find a currently existing AD flex
asset, which allows for the ability to manage multiple domains
automatically.

Specifically, the script looks for the FIRST match based on domain name
(which should be immutable for these flex assets) and picks that, so if
there are duplicate flex assets with the same domain, the one with the
lowest ID is going to be returned and modified.