Update for ITGlueAPI Module v2.0.0 Compliance
The ITGlueAPI PowerShell module was recently updated to version 2.0.0.
With that came a breaking change wherein the `data` key is no longer
explicitly required to be formed in the object tossed to any of the
`New-` or `Update-` functions.

This update allows the scripts to work with v2.0.0 of the module.